pyproject.toml
[project]
name = "datahub-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acryl-datahub[datahub-rest,postgres]>=1.5.0.13",
]
